+++ R/confint.cirls.R
@@ -13,15 +13,8 @@ confint.cirls <- function(object, parm, level = 0.95, nsim = 1000,
 
   dots <- list(...)
 
-  # Select coefficients
-  aliased <- stats::summary.glm(object)$aliased
-  pnames <- names(aliased)
-  if (missing(parm))
-    parm <- pnames
-  else if (is.numeric(parm))
-    parm <- pnames[parm]
-
   # Remove if aliased
+  aliased <- stats::summary.glm(object)$aliased
   if (!complete) parm <- parm[!aliased[parm]]
 
   # simulate from truncated multivariate normal
@@ -29,11 +22,5 @@ confint.cirls <- function(object, parm, level = 0.95, nsim = 1000,
   simures <- simulCoef(object, nsim = nsim, complete = TRUE, seed = seed)
 
   # Compute limits
-  lims <- c((1 - level) / 2, level + (1 - level) / 2)
-  res <- t(apply(simures[, parm, drop = F], 2, stats::quantile, lims,
-    na.rm = TRUE))
-  colnames(res) <- c("low", "high")
-
-  # Return
-  return(res)
+  confint.sim.cirls(simures, parm = parm, level = level)
 }